refactor: move misplaced integration/e2e tests out of tests/unit/
Moves tests that require network access, model downloads, or specific
hardware (NPU/GPU) from tests/unit/ to tests/integration/ or tests/e2e/,
and fixes misleading test file names.

Closes #51
