+++ tests/integration/datasets/test_object_detection.py
