+++ tests/integration/test_module_build.py
