+++ tests/unit/models/auto/test_automodel.py
